SR-2705 | Move things around
