+++ lib/associationProcessorService/index.js
